moved StaffListItem to component package
